Enhance documentation workflow in CodingAgent
Updated `CodingAgentProcess.cs` to introduce new classes for gathering product information, generating documentation, applying suggestions, publishing documentation, and proofreading documentation. Added `GeneratedDocumentationState` for managing chat history.

Refactored `CodingAgentEndpoints.cs` to remove unused imports and streamline endpoint mapping, focusing on essential parameters and simplifying task acknowledgment. Commented out previous orchestration logic to reflect the new workflow structure.
